telegram-notifier/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
paho-mqtt==1.6.1
python-telegram-bot==20.7
pydantic==1.10.14
